% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO_JH.R
\name{enrichGO_JH}
\alias{enrichGO_JH}
\title{Modified GO Enrichment Analysis from clusterprofiler enrichGO.}
\usage{
enrichGO_JH(gene, OrgDb, keyType = "ENTREZID", ont = "MF",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe,
  qvalueCutoff = 0.2, minGSSize = 10, maxGSSize = 500,
  readable = FALSE, pool = FALSE)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{OrgDb}{OrgDb}

\item{keyType}{keytype of input gene}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{qvalueCutoff}{qvalue cutoff}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{readable}{whether mapping gene ID to gene Name}

\item{pool}{If ont='ALL', whether pool 3 GO sub-ontologies}
}
\value{
An \code{enrichResult} instance.
}
\description{
This modified fucntion through pre-load GO_DATA("CC","MF","BP) 
into the enviroment to save time.
Given a vector of genes, this function will return the enrichment GO
categories after FDR control.
}
\examples{
\dontrun{
  data(geneList, package = "DOSE")
	de <- names(geneList)[1:100]
	yy <- enrichGO(de, 'org.Hs.eg.db', ont="BP", pvalueCutoff=0.01)
	head(yy)
}
# have to keep rds in some local folder
# "/stor/jianghao/Database/GO_DATA"
GO_DATA_BP <- readRDS("./GO_DATA/GO_DATA_BP.rds")
GO_DATA_CC <- readRDS("./GO_DATA/GO_DATA_CC.rds")
GO_DATA_MF <- readRDS("./GO_DATA/GO_DATA_MF.rds")

ego_CC <- enrichGO_JH(gene          = gene,
                     universe      = names(geneList),
                     OrgDb         = org.Hs.eg.db,
                     ont           = "CC", # chose ontology terms
                     pAdjustMethod = "BH", 
                     pvalueCutoff  = 0.01,
                     qvalueCutoff  = 0.05,
                     readable      = TRUE)
head(ego_BP)
                     
ego_CC <- enrichGO_JH(gene          = gene,
universe      = names(geneList),
OrgDb         = org.Hs.eg.db,
ont           = "BP", # chose ontology terms
pAdjustMethod = "BH", 
pvalueCutoff  = 0.01,
qvalueCutoff  = 0.05,
readable      = TRUE)                     
                     
                  
}
\seealso{
\code{\link{enrichResult-class}}, \code{\link{compareCluster}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\keyword{manip}
